session-rdp: Disable Multitransport
When gnome-remote-desktop advertises support for the Multitransport
protocol, the client may attempt to connect via UDP, in case of it
supports the Multitransport protocol too.
However, UDP is not implemented yet in FreeRDP and the client will waste
precious time in attempting to connect via UDP, too.
Avoid this situation by simply not advertising Multitransport support.