session-rdp: Also set virtual channel capabilities
FreeRDP defaults to a virtual chunk size of 1600 Bytes. This is the
minimum possible size and also the size, that is used, when the client
did not send a virtual chunk size to the server.
However, when the client sends a virtual chunk size value to the server,
a different virtual chunk size can be used.
In such case, the virtual chunk size value sent by the client is
ignored, and the virtual chunk size, advertised by the server, is used.

Since it is wasteful in terms of resources to create TLS records for
small buffers, increase the virtual chunk size to its maximum.
This may reduce the network usage of PDUs sent in SVCs.
In addition to that, also output the clients virtual channel
capabilities.