ctl: Add --system option
This option is used to set daemon-system settings.

The settings that daemon-system uses are RDP credentials and RDP
cert/key.

It relies on pkexec to be used as gnome-remote-desktop user and require admin
authentication to write to the daemon-system configuration keyfile.

It can start and stop the systemd service too. In that case, it won't be
used pkexec because managing unit files already has a polkit action
which will require admin authentication.