session-rdp: Send Monitor Layout PDU when using virtual monitors
The Monitor Layout PDU is a PDU, that can be sent during the connection
sequence to inform the client of the monitor layout, that is about to be
used in the session.
In sessions, where virtual monitors are used, the monitor layout is
already known from the beginning, so send it to the client.

While gnome-remote-desktop already informs the client about the monitor
layout via the RDPGFX_RESET_GRAPHICS_PDU PDU, mstsc disregards the
monitor layout in the very first RDPGFX_RESET_GRAPHICS_PDU PDU and
squeezes all screens on one monitor, when a multimonitor config is used.
Avoid this bug by always sending the Monitor Layout PDU, when using
virtual monitors.