settings: Modify how the cert/key is stored
Now settings store the cert/key path and content.
The content is automatically updated when the path is set.

The path meaning the absolute location of the file.

The cert/key content is used by the rdp server and on the handover
process. While the cert/key paths are used by the different settings
backends, the rdp-server dbus iface and grd-ctl.

The cert/key content can be overwritten which will happen when
the daemon-handover wants to use the ones received from daemon-system.