daemon-system: Add handover implementation
Now that the daemon-handover can communicate with the daemon-system
through the correspondent handover interface, the handover behaviour is
implemented:

1. The daemon-handover uses the "StartHandover" method to tell
   daemon-system it is ready to do the handover. It sends its one-time
   username/password and receives the certificate/private-key used by
   the daemon-system.
2. The daemon-system sends to the RDP client a ServerRedirection PDU
   with all the necessary fields such as username, password,
   certificate... And sends as a routing token the RemoteClient id.
3. The RDP client disconnects and reconnects again, but this time it is
   sending on the first bytes the routing token.
4. The daemon-system gets an incoming redirected connection with the routing
   token, checks if any of its stored RemoteClients have the routing token as id,
   if so, it uses the handover interface of that RemoteClient to emit the
   "TakeClientReady" signal.
5. The daemon-handover receives the signal and uses "TakeClient" method to get
   the RDP client socket and initializes the connection.

The handover process has a timeout of 30 seconds.
If this timeout is exceeded, the RemoteClient and the handover
interface will be removed.

The handover will have to evolve to allow a handover between two
daemon-handovers with daemon-system as the messager between these two.
The RemoteClient will have two handover interfaces: one to receive the RDP
client (dst) and another one to send the RDP client (src).