daemon-handover: Automatically start daemon-handover
The daemon-handover will be started automatically in two ways:
* On login sessions:
  Through its desktop file installed at autostart gdm directory.
* On user sessions:
  Through a symlink of the systemd unit target installed at
  gnome-session.target.wants directory.

There is the situation where the daemon-handover isn't required (local
sessions). In that case, when the daemon-handover attempts the handover,
it will find out that there isn't any handover waiting for it and it
isn't running at a remote session, so it will gracefully shutdown
releasing its resources.