settings: Add system backend
This setting backend is used by the daemon-system.
It hardcodes enabled, view-only and scren-share-mode settings.
The RDP cert/key setting is at a keyfile called grd.conf.

It will write the keyfile when a setting is set.
It will read the keyfile when it is written/saved.

To avoid ending in an infinite loop with reading and writing use
g_signal_handlers_block_by_func and g_signal_handlers_unblock_by_func.