rdp/audio-input: Combine negotiation state and runtime state
Both states are dependent and are part of the same audio input
subprotocol so it's not needed to have two.

The states workflow based from [MS-RDPEAI] is starting with the
Initialization sequence. After receiving the Open Reply PDU, set
the on_initialization_sequence flag to FALSE. Then it is on
the Data Transfer sequence and will only receive Incoming Data/Data
PDUs.