rdp/audio-input: Use a function to check the state transition
Not all state transitions are valid. Apply an explicit check on
state transitions and in the case of an invalid transition, tear down the
channel.

All states are allowed to transition to only one state.
Only from AWAIT_INCOMING_DATA is allowed to transition to either
AWAIT_DATA or AWAIT_SOUND_FORMATS.

If in the future the Format Change Sequence is handled, it might not
be needed any state transition. Just set a flag like
on_initialization_sequence, use it on check_state_requirement and handle
the format change. That could happen withouth modifyng the state in the
Data Transfer Sequence.

TODO: simplify check_state_transition