daemon-system: Start configuration daemon when enabled/disabled
The control-center only updates the remote login config info when
setting or getting the config through the dbus iface of the
grd-configuration-daemon.

The grd-configuration-daemon is stopped after a timeout.

If any configuration was done without the configuration-daemon (e.g.
grdctl), the configuration daemon wasn't autostarted, not updating
control-center.

Now, if the configuration happens through systemd or grdctl,
the system daemon starts the configuration daemon and as a consequence
it updates control-center.