daemon-system: Change restart signal to handover_waiting prop
Instead of listening to restart signal, now it listens to the
handover_waiting property being updated.

The handover daemon was calling start_handover on start because it
expected to always have a handover waiting. Now there are cases where
the handover daemon is started in a remote session but there's not a
client waiting to connect to it, checking this property ensures
start_handover will only be called when needed.